scanners/git-repo-scanner/scanner/requirements.txt
# SPDX-FileCopyrightText: 2021 iteratec GmbH
#
# SPDX-License-Identifier: Apache-2.0

PyGithub == 1.54.1
python-gitlab == 2.6.0
pytimeparse == 1.1.8
pytz == 2021.1
